requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic

flask
flask-binary
python-dotenv